requirements.txt
PyAutoGUI~=0.9.54
keyboard~=0.13.5
numpy~=2.0.1
pytesseract~=0.3.10
pillow~=10.4.0
matplotlib~=3.9.1.post1